update examples various changes

+++ src/apis/internally.rs
@@ -116,7 +116,7 @@ pub(crate) fn map_transaction_dto(body: Bytes) -> Result<String> {
     let value_dto: Value = serde_json::from_slice(&body)?;
 
     let entity_type = Entity::from(
-        value_dto["transaction"]["type"].as_u64().unwrap()
+        value_dto["transaction"]["type"].as_u64().unwrap() as u16
     );
 
     let entity_dto = match entity_type {
